+++ lib/src/core/components/index.tsx
@@ -28,3 +28,6 @@ export * from "./CenteredView";
 export * from "./InfoLabel";
 
 export * from "./fields/FilledSelect";
+export * from "./fields/TextInput";
+export * from "./fields/TextareaAutosize";
+export * from "./fields/DisabledTextField";